agent_cli/requirements.txt
requests>=2.20.0        # For test client login
python-socketio[asyncio_client]>=5.8.0 # For WebSocket client
aioconsole>=0.4.0
